OSINTBABY/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
dnspython>=2.3.0
python-whois>=0.8.0
click>=8.1.0
rich>=13.0.0
aiohttp>=3.8.0
pyyaml>=6.0
validators>=0.20.0
shodan>=1.28.0